django/requirements.txt
Django==5.0.3
djangorestframework==3.14.0
uWSGI==2.0.22
django-environ==0.11.2
django-cors-headers==4.4.0